WxWidgets 3.1 adaptation
Several deprecated functions and other issues havd to be updated.
(SetCurrent wxGLCanvas, MakeModal (wxFrame) wxGrid... and so on...
A small bug in ApoloMex CMakeList (FinMatlab.cmake)  has been fixed too

+++ connectionLog.h
@@ -27,7 +27,7 @@ private:
 	vector<NodeTree*>log;
 	wxString tittle;
 	wxWindow *parent;
-	double row,col;
+	int row,col;
 	Mutex mutex;
